+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [Unreleased]
 ### Added
 - ext config stores can control 'ssl' flag of underlying store with 'scheme' attribute
+- env fallback reference
 
 ### Changed
 - ServiceClientFactory deprecated by VertxEndpointClient